Invoicing logic was improved, deduplication of code
